cube-ai-agent/requirements.txt
fastapi
uvicorn
requests
python-dotenv
requests-mock